requirements.txt
# ============================================================
# Escucha — pinned production dependencies
# ============================================================
# INSTALL ORDER MATTERS on Windows with GPU:
#
#   Step 1 — Install PyTorch with CUDA 12.8 first:
#     pip install torch==2.11.0 torchaudio==2.11.0 --index-url https://download.pytorch.org/whl/cu128
#
#   Step 2 — Install everything else:
#     pip install -r requirements.txt
#
#   CPU-ONLY alternative for Step 1:
#     pip install torch==2.11.0 torchaudio==2.11.0 --index-url https://download.pytorch.org/whl/cpu
#
# ============================================================

# -- Web server -----------------------------------------------
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.20

# -- Transcription --------------------------------------------
faster-whisper==1.2.1

# -- Diarization ----------------------------------------------
pyannote-audio==4.0.4

# -- Summarization --------------------------------------------
httpx==0.28.1
anthropic==0.52.0

# -- Utilities ------------------------------------------------
python-dotenv==1.1.0
pydantic==2.13.3
soundfile>=0.12.1                 # WAV loader; bundles libsndfile DLL —
                                  # used by diarizer to bypass torchcodec/FFmpeg-DLL deps
